requirements.txt
icloudpd
pyexiftool